requirements.txt
flask
transformers
torch
pdfminer.six
nltk
pandas
streamlit
